+++ Sources/DocumentKit/DocumentGroupModal.swift
@@ -11,16 +11,42 @@ import SwiftUI
 /**
  This protocol can be implemented by any view, that you have
  to be able to open from a document group.
- */
-public protocol DocumentGroupModal: View, DocumentGroupInspector {}
 
-public extension DocumentGroupModal {
+ Note that these presentation functions are available to all
+ SwiftUI views, since you must be able to present views with
+ applied view modifiers.
+ */
+public protocol DocumentGroupModal: View, DocumentGroupInspector {
+
+    /**
+     Present the view as a document group sheet.
+     */
+    func presentAsDocumentGroupSheet() throws
+
+    /**
+     Present the view as a document group full screen cover.
+     */
+    func presentAsDocumentGroupFullScreenCover() throws
+
+    /**
+     Present the view as a document group modal.
+     */
+    func presentAsDocumentGroupModal(_ style: UIModalPresentationStyle ) throws
+}
+
+/// This internal inspector is used by the view extensions.
+private class InternalInspector: DocumentGroupInspector {
+
+    static var shared = InternalInspector()
+}
+
+public extension View {
 
     /**
      Present the view as a document group sheet.
      */
     func presentAsDocumentGroupSheet() throws {
-        try presentAsDocumentGroupModal()
+        try presentAsDocumentGroupModal(.automatic)
     }
 
     /**
@@ -33,12 +59,11 @@ public extension DocumentGroupModal {
     /**
      Present the view as a document group modal.
      */
-    func presentAsDocumentGroupModal(
-        _ presentationStyle: UIModalPresentationStyle = .automatic
-    ) throws {
-        guard let parent = rootViewController else { throw DocumentGroupError.noParentWindow }
+    func presentAsDocumentGroupModal(_ style: UIModalPresentationStyle) throws {
+        let inspector = InternalInspector.shared
+        guard let parent = inspector.rootViewController else { throw DocumentGroupError.noParentWindow }
         let controller = UIHostingController(rootView: self)
-        controller.modalPresentationStyle = presentationStyle
+        controller.modalPresentationStyle = style
         parent.present(controller, animated: true, completion: nil)
     }
 }